change upgrade scripts

+++ source/gqlcnode_upgrade/preinstall.sh
@@ -7,11 +7,17 @@ fi
  
 killall gqlc-confidant
 #check qlc_update.py
-diff /tmp/gqlcnode_upgrade/qlc_update.py /root/qlc_update.py >/dev/null
-if  [ $0 != 0 ];then
-	cp -f /tmp/gqlcnode_upgrade/qlc_update.py /root/qlc_update.py
+if [ -f /tmp/gqlcnode_upgrade/qlc_update.py ];then
+	diff /tmp/gqlcnode_upgrade/qlc_update.py /root/qlc_update.py >/dev/null
+	if  [ $0 != 0 ];then
+		cp -f /tmp/gqlcnode_upgrade/qlc_update.py /root/qlc_update.py
+	fi
 fi
 
 #删除原有账本目录
-rm -f nohup.out
-rm -rf /sata/home/gqlcnode/*  
+if [ -f /root/nohup.out ];then
+	rm -f nohup.out
+fi
+if [ -d /sata/home/gqlcnode ];then
+	rm -rf /sata/home/gqlcnode/*
+fi
